requirements.txt
beautifulsoup4==4.10.0
pyTelegramBotAPI==4.2.2
python-dotenv==0.19.2
requests==2.27.0
